Add profile to connect to bluetooth device
This has been tested with the A2DP and Headset profiles on two different
bluetooth devices that I had at hand, but I didn't have easy access to the other
device profiles. They should work the same though.

There doesn't seem to be an officially supported method of automating connecting
to already paired devices, but this method using reflection seems to be quite
stable, as the code I imported (Apache 2.0 License) is from 2013, but still
works (tested with Android Pie).